Rename useI18n to useLocaleText
