+++ packages/grid/_modules_/grid/constants/index.ts
@@ -1,4 +1,4 @@
 export * from './envConstants';
 export * from './eventsConstants';
 export * from './cssClassesConstants';
-export * from './i18nConstants';
+export * from './localeTextConstants';
