--- packages/grid/_modules_/grid/hooks/features/i18n/index.ts
@@ -1,1 +1,0 @@
-export * from './useI18n';
