+++ packages/grid/_modules_/grid/hooks/features/localeText/index.ts
@@ -1,0 +1,1 @@
+export * from './useLocaleText';
