+++ packages/grid/_modules_/grid/models/api/index.ts
@@ -10,4 +10,4 @@ export * from './selectionApi';
 export * from './sortApi';
 export * from './stateApi';
 export * from './virtualizationApi';
-export * from './i18nApi';
+export * from './localeTextApi';
