+++ packages/grid/_modules_/grid/models/api/localeTextApi.ts
@@ -5,9 +5,9 @@ export type LocaleTextValue = string | React.ReactNode | Function;
 export type TranslationKeys = keyof LocaleText;
 
 /**
- * The i18n API interface that is available in the grid [[apiRef]].
+ * The grid localeText API [[apiRef]].
  */
-export interface I18nApi {
+export interface LocaleTextApi {
   /**
    * Get grid text.
    * @param key
